refactor: move common info into MediaInfo
