+++ src/parser/youtube/VideoInfo.ts
@@ -1,6 +1,3 @@
-import Constants from '../../utils/Constants.js';
-import Parser from '../index.js';
-
 import ChipCloud from '../classes/ChipCloud.js';
 import ChipCloudChip from '../classes/ChipCloudChip.js';
 import CommentsEntryPointHeader from '../classes/comments/CommentsEntryPointHeader.js';
@@ -24,7 +21,6 @@ import NavigationEndpoint from '../classes/NavigationEndpoint.js';
 
 import type CardCollection from '../classes/CardCollection.js';
 import type Endscreen from '../classes/Endscreen.js';
-import type Format from '../classes/misc/Format.js';
 import type PlayerAnnotationsExpanded from '../classes/PlayerAnnotationsExpanded.js';
 import type PlayerCaptionsTracklist from '../classes/PlayerCaptionsTracklist.js';
 import type PlayerLiveStoryboardSpec from '../classes/PlayerLiveStoryboardSpec.js';
@@ -32,32 +28,20 @@ import type PlayerStoryboardSpec from '../classes/PlayerStoryboardSpec.js';
 
 import type Actions from '../../core/Actions.js';
 import type { ApiResponse } from '../../core/Actions.js';
-import type Player from '../../core/Player.js';
-import type { ObservedArray, YTNode } from '../helpers.js';
-import type { INextResponse, IPlayerResponse } from '../types/ParsedResponse.js';
-
-import FormatUtils, { DownloadOptions, FormatFilter, FormatOptions, URLTransformer } from '../../utils/FormatUtils.js';
+import { ObservedArray, YTNode } from '../helpers.js';
 
 import { InnertubeError } from '../../utils/Utils.js';
+import { MediaInfo } from '../../core/MediaInfo.js';
 
-class VideoInfo {
-  #page: [IPlayerResponse, INextResponse?];
-
-  #actions: Actions;
-  #player?: Player;
-  #cpn?: string;
+class VideoInfo extends MediaInfo {
   #watch_next_continuation?: ContinuationItem;
 
   basic_info;
-  streaming_data;
-  playability_status;
   annotations?: ObservedArray<PlayerAnnotationsExpanded>;
   storyboards?: PlayerStoryboardSpec | PlayerLiveStoryboardSpec;
   endscreen?: Endscreen;
   captions?: PlayerCaptionsTracklist;
   cards?: CardCollection;
-
-  #playback_tracking;
 
   primary_info?: VideoPrimaryInfo | null;
   secondary_info?: VideoSecondaryInfo | null;
@@ -77,18 +61,10 @@ class VideoInfo {
    * @param player - Player instance.
    * @param cpn - Client Playback Nonce.
    */
-  constructor(data: [ApiResponse, ApiResponse?], actions: Actions, player?: Player, cpn?: string) {
-    this.#actions = actions;
-    this.#player = player;
-    this.#cpn = cpn;
+  constructor(data: [ApiResponse, ApiResponse?], actions: Actions, cpn: string) {
+    super(data, actions, cpn);
 
-    const info = Parser.parseResponse<IPlayerResponse>(data[0].data);
-    const next = data?.[1]?.data ? Parser.parseResponse<INextResponse>(data[1].data) : undefined;
-
-    this.#page = [ info, next ];
-
-    if (info.playability_status?.status === 'ERROR')
-      throw new InnertubeError('This video is unavailable', info.playability_status);
+    const [ info, next ] = this.page;
 
     if (info.microformat && !info.microformat?.is(PlayerMicroformat, MicroformatData))
       throw new InnertubeError('Invalid microformat', info.microformat);
@@ -113,15 +89,11 @@ class VideoInfo {
       is_disliked: undefined as boolean | undefined
     };
 
-    this.streaming_data = info.streaming_data;
-    this.playability_status = info.playability_status;
     this.annotations = info.annotations;
     this.storyboards = info.storyboards;
     this.endscreen = info.endscreen;
     this.captions = info.captions;
     this.cards = info.cards;
-
-    this.#playback_tracking = info.playback_tracking;
 
     const two_col = next?.contents?.item().as(TwoColumnWatchNextResults);
 
@@ -199,7 +171,7 @@ class VideoInfo {
 
     if (cloud_chip.is_selected) return this;
 
-    const response = await cloud_chip.endpoint?.call(this.#actions, { parse: true });
+    const response = await cloud_chip.endpoint?.call(this.actions, { parse: true });
     const data = response?.on_response_received_endpoints?.get({ target_id: 'watch-next-feed' });
 
     this.watch_next_feed = data?.contents;
@@ -211,24 +183,7 @@ class VideoInfo {
    * Adds video to the watch history.
    */
   async addToWatchHistory(): Promise<Response> {
-    if (!this.#playback_tracking)
-      throw new InnertubeError('Playback tracking not available');
-
-    const url_params = {
-      cpn: this.#cpn,
-      fmt: 251,
-      rtn: 0,
-      rt: 0
-    };
-
-    const url = this.#playback_tracking.videostats_playback_url.replace('https://s.', 'https://www.');
-
-    const response = await this.#actions.stats(url, {
-      client_name: Constants.CLIENTS.WEB.NAME,
-      client_version: Constants.CLIENTS.WEB.VERSION
-    }, url_params);
-
-    return response;
+    return super.addToWatchHistory();
   }
 
 
@@ -239,7 +194,7 @@ class VideoInfo {
     if (!this.#watch_next_continuation)
       throw new InnertubeError('Watch next feed continuation not found');
 
-    const response = await this.#watch_next_continuation?.endpoint.call(this.#actions, { parse: true });
+    const response = await this.#watch_next_continuation?.endpoint.call(this.actions, { parse: true });
     const data = response?.on_response_received_endpoints?.get({ type: 'appendContinuationItemsAction' });
 
     if (!data)
@@ -271,7 +226,7 @@ class VideoInfo {
     if (button.is_toggled)
       throw new InnertubeError('This video is already liked', { video_id: this.basic_info.id });
 
-    const response = await button.endpoint.call(this.#actions);
+    const response = await button.endpoint.call(this.actions);
 
     return response;
   }
@@ -292,7 +247,7 @@ class VideoInfo {
     if (button.is_toggled)
       throw new InnertubeError('This video is already disliked', { video_id: this.basic_info.id });
 
-    const response = await button.endpoint.call(this.#actions);
+    const response = await button.endpoint.call(this.actions);
 
     return response;
   }
@@ -320,7 +275,7 @@ class VideoInfo {
     if (!button)
       throw new InnertubeError('This video is not liked/disliked', { video_id: this.basic_info.id });
 
-    const response = await button.toggled_endpoint.call(this.#actions);
+    const response = await button.toggled_endpoint.call(this.actions);
 
     return response;
   }
@@ -335,50 +290,10 @@ class VideoInfo {
   }
 
   /**
-   * Selects the format that best matches the given options.
-   * @param options - Options
-   */
-  chooseFormat(options: FormatOptions): Format {
-    return FormatUtils.chooseFormat(options, this.streaming_data);
-  }
-
-  /**
-   * Generates a DASH manifest from the streaming data.
-   * @param url_transformer - Function to transform the URLs.
-   * @param format_filter - Function to filter the formats.
-   * @returns DASH manifest
-   */
-  toDash(url_transformer?: URLTransformer, format_filter?: FormatFilter): string {
-    return FormatUtils.toDash(this.streaming_data, url_transformer, format_filter, this.#cpn, this.#player);
-  }
-
-  /**
-   * Downloads the video.
-   * @param options - Download options.
-   */
-  async download(options: DownloadOptions = {}): Promise<ReadableStream<Uint8Array>> {
-    return FormatUtils.download(options, this.#actions, this.playability_status, this.streaming_data, this.#actions.session.player, this.cpn);
-  }
-
-  /**
    * Watch next feed filters.
    */
   get filters(): string[] {
     return this.related_chip_cloud?.chips?.map((chip) => chip.text?.toString()) || [];
-  }
-
-  /**
-   * Actions instance.
-   */
-  get actions(): Actions {
-    return this.#actions;
-  }
-
-  /**
-   * Content Playback Nonce.
-   */
-  get cpn(): string | undefined {
-    return this.#cpn;
   }
 
   /**
@@ -432,13 +347,6 @@ class VideoInfo {
         return songs;
         */
     return [];
-  }
-
-  /**
-   * Original parsed InnerTube response.
-   */
-  get page(): [IPlayerResponse, INextResponse?] {
-    return this.#page;
   }
 }
